requirements.txt
annotated-types
anyio
blinker
certifi
charset-normalizer
click
fastapi
Flask
h11
httpcore
httpx
idna
itsdangerous
Jinja2
MarkupSafe
pydantic
pydantic_core
python-dotenv
pytube
redis
requests
six
sniffio
spotipy
starlette
typing_extensions
urllib3
uvicorn
Werkzeug
